add OpenIdAuthorizer into configuation file
