+++ server/src/assembly/resources/conf/iotdb-engine.properties
@@ -479,3 +479,18 @@ mqtt_handler_pool_size=1
 
 # the mqtt message payload formatter.
 mqtt_payload_formatter=json
+
+
+####################
+### Authorization Configuration
+####################
+
+#which class to serve for authorization. By default, it is LocalFileAuthorizer.
+#Another choice is org.apache.iotdb.db.auth.authorizer.OpenIdAuthorizer
+authorizer_provider_class=org.apache.iotdb.db.auth.authorizer.LocalFileAuthorizer
+
+
+#If OpenIdAuthorizer is enabled, then openID_url must be set.
+
+#openID_url=
+
